structure model computing files
